Convert image.rb to a override file and remove the unused imagemagick code
